Keystone Project Restructuring
- based on the initial need of Optimized Docker Builds
