+++ cms/admin/pages/dashboard/list.tsx
@@ -1,62 +1,64 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 
-import { HTMLAttributes, ReactNode } from 'react';
-import { jsx, useTheme } from '@keystone-ui/core';
-  
-  export const TableContainer = ({ children }: { children: ReactNode }) => {
-    return (
-      <table
-        css={{
-          minWidth: '100%',
-          tableLayout: 'fixed',
-          'tr:last-child td': { borderBottomWidth: 0 },
-        }}
-        cellPadding="0"
-        cellSpacing="0"
-      >
-        {children}
-      </table>
-    );
-  };
-  
-  export const TableHeaderRow = ({ children }: { children: ReactNode }) => {
-    return (
-      <thead>
-        <tr>{children}</tr>
-      </thead>
-    );
-  };
-  
-  export const TableHeaderCell = (props: HTMLAttributes<HTMLElement>) => {
-    const { colors, spacing, typography } = useTheme();
-    return (
-      <th
-        css={{
-          backgroundColor: colors.background,
-          borderBottom: `2px solid ${colors.border}`,
-          color: colors.foregroundDim,
-          fontSize: typography.fontSize.medium,
-          fontWeight: typography.fontWeight.medium,
-          padding: spacing.small,
-          textAlign: 'left',
-          position: 'sticky',
-          top: 0,
-        }}
-        {...props}
-      />
-    );
-  };
-  
-  export const TableBodyCell = (props: HTMLAttributes<HTMLElement>) => {
-    const { colors, typography } = useTheme();
-    return (
-      <td
-        css={{
-          borderBottom: `1px solid ${colors.border}`,
-          fontSize: typography.fontSize.medium,
-        }}
-        {...props}
-      />
-    );
-  };
+import { HTMLAttributes, ReactNode } from "react";
+import { jsx, useTheme } from "@keystone-ui/core";
+
+export const TableContainer = ({ children }: { children: ReactNode }) => {
+  return (
+    <table
+      css={{
+        minWidth: "100%",
+        tableLayout: "fixed",
+        "tr:last-child td": { borderBottomWidth: 0 },
+      }}
+      cellPadding="0"
+      cellSpacing="0"
+    >
+      {children}
+    </table>
+  );
+};
+
+export const TableHeaderRow = ({ children }: { children: ReactNode }) => {
+  return (
+    <thead>
+      <tr>{children}</tr>
+    </thead>
+  );
+};
+
+export const TableHeaderCell = (props: HTMLAttributes<HTMLElement>) => {
+  const { colors, spacing, typography } = useTheme();
+  return (
+    <th
+      css={{
+        backgroundColor: colors.background,
+        borderBottom: `2px solid ${colors.border}`,
+        color: colors.foregroundDim,
+        fontSize: typography.fontSize.medium,
+        fontWeight: typography.fontWeight.medium,
+        padding: spacing.small,
+        textAlign: "left",
+        position: "sticky",
+        top: 0,
+      }}
+      {...props}
+    />
+  );
+};
+
+export const TableBodyCell = (props: HTMLAttributes<HTMLElement>) => {
+  const { colors, typography } = useTheme();
+  return (
+    <td
+      css={{
+        borderBottom: `1px solid ${colors.border}`,
+        fontSize: typography.fontSize.medium,
+      }}
+      {...props}
+    />
+  );
+};
+
+export default TableContainer;
